Fix source PR identification for `push` events (#26)
`github.event.number` is only populated for actions triggered by pull
request `event`s, not `push` etc. As we use this to tell `gh` on which
PR to put the comments, when it's empty `gh` simply uses the new
backported PR as a fallback.

GitHub doesn't give the action the source PR on a `push` event, but [we
can use the GitHub API to query the PR of the commit that was `push`ed
to figure it
out](https://docs.github.com/en/rest/commits/commits?apiVersion=2022-11-28#list-pull-requests-associated-with-a-commit),
and instead use that in those cases.

Testing:
- [PR backported by label added post-merge (e.g.
`hazelcast-mono`)](https://github.com/JackPGreen/backport-test/pull/137)
- [PR backported by `push` trigger (e.g.
`hz-docs`)](https://github.com/JackPGreen/backport-test/pull/139)

Fixes: https://github.com/hazelcast/backport/issues/25, [Slack
discussion](https://hazelcast.slack.com/archives/C07066ELRRD/p1743505394712119)