Retain Jira references in backport titles (#30)
The script removes any previous suffixes when computing new commit title
(e.g. when backporting `blah [5.5.z]` to `5.4.z`, the `5.5.z` references
should be removed. But this check is too fuzzy and `[anything]` is
removed - which is a problem when Jira references are used (and
subsequently, removed).

It's not possible to know what format the previous suffix _might've_
been (it could be any valid branch name), but to workaround this
_specific_ issue we can check it doesn't contain a `-` - in our typical
use-case, this is enough to differentiate a branch from a Jira issue
reference.

Fixes: [DI-616](https://hazelcast.atlassian.net/browse/DI-616)

[DI-616]:
https://hazelcast.atlassian.net/browse/DI-616?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ