Support backporting before _and_ after PR is merged (#31)
The existing backport action was _primarily_ used in conjunction with
[`hazelcast-mono/backport.yml`](https://github.com/hazelcast/hazelcast-mono/blob/master/.github/workflows/backport.yml).

Between them:
- when a PR was labelled
- if it was merged
- it was backported

That workflow works fine, but it means a PR cannot be _pre-emptively_
labelled a backport before it's merged - this means in practice, for a
trivial change you have to set a post-merge reminder to come back and
add the label. This was an implementation limitation - the action was
triggered when the PR was labelled, and the GitHub `event` contained the
triggering label.

It would be preferable if instead you could label a PR at _any_ time,
and only once merged will it backported (or instantaneously, if already
so).

`hz-docs` has similar backporting functionality that works in reverse -
and this is opportunity to merge both.

Here there are two primary architectural changes:
1. Support un-merged PRs
- using a [special
action](https://github.com/marketplace/actions/get-current-pull-request),
we can get the associated PR (and thus labels) from a `push` event
- we can then evaluate and execute required backports for those label(s)
2. Merge into a single action
- half of the implementation was in this repo, the other half was
implemented differently between `hazelcast-mono` & `hz-docs`
- instead, have a single (simply) entry point to trigger backporting on
labels

Additional changes:
- When reacting to a `labelled` event, it was implicit only a single
event would be sent at once - but now we need to support multiple labels
on a single PR
- each backport-identified label is removed from the PR (to prevent some
kind of backport->forwardport storm)
   - each label is backported in parallel via a matrix
- we also have to address concurrency concerns
   - previously this was ignored
- but now we use a lock to ensure an individual PR is not being
processed in multiple runners concurrently
- runners changed from `ubicloud` to GitHub runners
   - because _some_ callers will be public repos (where usage is free)
- to reduce costs, using [GitHub's new `ubuntu-slim`
runner](https://github.blog/changelog/2025-10-28-1-vcpu-linux-runner-now-available-in-github-actions-in-public-preview/)
 - `omit-labels` functionality removed
- this was added to workaround some of the limitations discussed above
that are now addressed -
https://github.com/hazelcast/backport/commit/013f934d5516282ad9a03622151ac8475de6e910

In https://github.com/hazelcast/hazelcast-mono/pull/3179 we were able to
optimise action invocation by skipping the entire job for any
non-backport labelling events to reduce costs. Now that we need to do
logic to figure out the labels, this is no longer possible, which means
_some_ usage every time a PR is labelled or a `push` is performed.

For testing, I've used a [test
repo](https://github.com/JackPGreen/backport-test) which can be used as
a sandbox to execute backports.

Fixes: [DI-685](https://hazelcast.atlassian.net/browse/DI-685)

[DI-685]:
https://hazelcast.atlassian.net/browse/DI-685?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ